Upgrade nom from version 2 to version 8
In nom version 4, the `IResult` type changed to alias the standard
library `Result`.

In nom version 5, there were a big changes. It was converted to
primarily use functions instead of macros. And now there are different
sets of functions for streaming vs complete input.

The macros use the streaming functions, which does not work for how they
are used in Bart because `is_not!()` won't match the remaining input
after the last tag. So the parsing is now converted to use functions
instead of macros and use `is_not` from nom's `complete` module.